% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_save_figures.R
\name{ks_save_figures}
\alias{ks_save_figures}
\title{Save Knowledge Sharing Figures}
\usage{
ks_save_figures(ks_result, dir = "ks_figures", width = 8, height = 4.5)
}
\arguments{
\item{ks_result}{Output of \code{\link[=run_ks]{run_ks()}}.}

\item{dir}{Directory to save into. Default: "ks_figures".}

\item{width, height}{Plot dimensions in inches.}
}
\description{
Save Knowledge Sharing Figures
}
