% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ks_plot.R
\name{visualize_ks_plot}
\alias{visualize_ks_plot}
\title{Visualize Knowledge Sharing (Windowed Centroid Change)}
\usage{
visualize_ks_plot(change_df, title_prefix = "")
}
\arguments{
\item{change_df}{A data frame as returned by \code{\link[=compute_centroid_change]{compute_centroid_change()}},
containing \code{anchor_time}, \code{cosine_to_prev}, \code{delta_euclid}.}

\item{title_prefix}{Character to prepend to plot titles. Default: "".}
}
\value{
A list of ggplot objects: \code{cosine_plot}, \code{euclid_plot}, \code{combined_plot}, and \code{cosine_colored_by_euclid_plot}.
}
\description{
Create line plots showing how much the centroid shifts between adjacent windows.
}
\examples{
\dontrun{
vis <- visualize_ks_plot(change_df)
vis$cosine_plot
}
}
